1) Creating PDB for release build
Branch created
CMakeList: special handling for 'none' targetDevice - treat is as missing
CMakeList: special handling for 'none' targetDevice - treat is as missing; do not set detected_target to 'none'
CMakeList: special handling for 'none' targetDevice - treat is as missing; review fix
ENV-132: Ability to add QT PDB files to builds
ENV-132: Ability to add QT PDB files to builds: review fixes
Fix license error messages in plural form
Fixed lanuching Lite Client on Nx1 for cmake: liteMode is now on.
Functional tests, scalability test: retry check requests 10 times, make longer timeout
Functional tests, scalability test: retry check requests 5 times, timeout 2min
Functional tests: add support for server 3.0 - check several .so libraries for cloud host tag
Functional tests: add support for server 3.0 - check several .so libraries for cloud host tag; review fixes
Functional tests: allow any host in account activation email
Functional tests: use separate email alias for every test group and customization; automatically create and activate email - review fixes
Implemented resource filtering and wxs generation
MOBILE-505: Video from RTSP is played with jerks on Android
Moved CustomActions to Cmake and fixed codestyle
NXTOOL-132 - macosx compilation fixes
NXTOOL-133 - Merge selected servers if both 'System Name' and 'Password' fields have input
Review Fix - 99 symbols
Review Fixes
Review Fixes #2
Review Fixes and messy properties refactoring
Review fix
Update translation strings
Update translations
VMS-5562 Incorrect error message when NICs can't be detected
VMS-6201: Move timeline by mouse wheel w/ Shift (in addition to Ctrl)
VMS-6271: new camera added to onvifPresets list
VMS-6300: Adapt distro building scripts for Linux, Nx1 and ISD to CMake
VMS-6444: Wrong speed on timeline zoom
VMS-6484 - Change order of presets
VMS-6720: Password stays filled in Login Dialog
VMS-6832: Item overlays and top panel have paddings from the edges
VMS-6854 Camera MAC isn't updated correctly
VMS-6876 - client crash
VMS-6876 - remove empty string
VMS-6924: If user have closed alarm layout, all cameras must be removed.
VMS-6928: Allow admin to create custom user's new layout via his 'Layouts' node
VMS-6943: Cannot load video in full-screen on Lite Client from all cameras on another server in merged system
VMS-6949: Export to nov. Video stops.
VMS-6972 - Invisible tiles on Welcome Screen
VMS-6973: Can't open advanced settings for ISD camera
VMS-6998: Bottom navigation panel arrow isn't hidden in full screen mode after 3 sec
VMS-7000 - Mac OS X. Mouse cursor hover stops working if camera was opened on the Scene by double click
VMS-7004: Incorrect stream resolution is set up on ISD cam
VMS-7010: Improve customizability of resource list view
VMS-7016: Wrong icon in 'remove items from layout' message box
VMS-7027: No stream URL's are shown in camera's settings for Hikvision DS-2CD3145F-I (cam with h265 support)
VMS-7030: Min/max available fps rates for camera Hikvision DS-2CD3145F-I (camera with h265 support) do not correspond to camera's capabilities
VMS-7031 - Welcome Screen. Server IP is updated on tile only after Client restart
VMS-7036 - Event Rules window can be resized on MacOs to minimum width (similar as it does on Windows)
VMS-7037: Wrong chunks selection
VMS-7038|Latest encoders added to the list
VMS-7047: NO_DATA label sometimes appears outside the item for a moment
VMS-7048: NO_DATA label is absent on IO_module without any chunks on it.
VMS-7051: Axis p7210 encoder can be added only as onvif resource
VMS-7052: Server requests motion detection for 4 sensors from AV10655 which has only 2 sensors.
VMS-7055: Fix file watermark
VMS-7055: Fix watermark for exported files
VMS-7056: Low quality is speeded up in some nov files
VMS-7060: 'Owner' role shouldn't be present for LDAP users roles list.
VMS-7064 - Soft trigger button is active and clickable when it is disabled by schedule
VMS-7064 - review fix
VMS-7066: Remove all mentions of 'Select Server to synchronize time with'
VMS-7067: AIO Thread crash
VMS-7068: Fix 'Push my screen' after relogin
VMS-7069 Allow Hardware ID 4 calculated without NIC
VMS-7073 - Notifications with Acknowledgement disappears by timeout and can be closed by 'x'
VMS-7084: Use licensing.networkoptix.com instead of nxlicensed.hdw.mx
VMS-7085: Mediaserver was crashed on Win7x64 (build 15852)
VMS-7089 - Nx Client connection dialog keeps password for the last used connection
VMS-7089 - cs fix
VMS-7089 - review fixes
VMS-7089: Fix 'Autologin' checkbox always set
VMS-7089: Removing saved connection now properly clears saved password
VMS-7098: extra emit 'timeChanged' on server startup
VMS-7099: Fix server 3.0 installer check
VMS-7103: License warning was immediatelly hidden after display in some circumstances
VMS-7105: Server restarts all the time after reboot
VMS-7107: Add 'Make Showreel' to layout context menu in tree
build testcamera, create assemblies, fix review feedback and use CMAKE_BUILD_TYPE
customactions build on cmake
duplicate 'Create Account' button in landing page for RAS. Fixed
edge1: Fixed autostart; added .flag-file-triggered logging to the autostart script.
installer.target.dir specifies target folder for installers
reverted accidentally committed code
review fix
version bump for beta
windows installer works, added missing modules, some review fixes

<==== Mercurial data ====>
Branch: vms_3.1_web

+++ edge_firmware/rpi/maven/filter-resources/install.sh
@@ -1,23 +1,39 @@
 #!/bin/bash
-set -x #< Log each command.
-set -u #< Prohibit undefined variables.
 set -e #< Exit on any error.
+set -u #< Prohibit undefined variables.
+
+configure()
+{
+    BOX="@box@"
+    FAILURE_FLAG="/var/log/vms-upgrade-failed.flag"
+    CUSTOMIZATION="@deb.customization.company.name@"
+    INSTALL_PATH="opt/$CUSTOMIZATION"
+    MEDIASERVER_PATH="$INSTALL_PATH/mediaserver"
+    LITE_CLIENT_PATH="$INSTALL_PATH/lite_client"
+    DISTRIB="@artifact.name.server@"
+    STARTUP_SCRIPT="/etc/init.d/$CUSTOMIZATION-mediaserver"
+    TAR_FILE="./$DISTRIB.tar.gz"
+}
+
+checkRunningUnderRoot()
+{
+    if [ "$(id -u)" != "0" ]; then
+        echo "ERROR: $0 should be run under root"
+        exit 1
+    fi
+}
 
 # Redirect all output of this script to the log file.
-LOG_FILE="/var/log/bpi-upgrade.log"
-exec 1<&- #< Close stdout fd.
-exec 2<&- #< Close stderr fd.
-exec 1<>"$LOG_FILE" #< Open stdout as $LOG_FILE for reading and writing.
-exec 2>&1 #< Redirect stderr to stdout.
+redirectOutput() # log_file
+{
+    LOG_FILE="$1"
+    echo "$0: All further output goes to $LOG_FILE"
 
-BOX="@box@"
-FAILURE_FLAG="/var/log/bpi-upgrade-failed.flag"
-CUSTOMIZATION="@deb.customization.company.name@"
-INSTALL_PATH="opt/$CUSTOMIZATION"
-MEDIASERVER_PATH="$INSTALL_PATH/mediaserver"
-DISTRIB="@artifact.name.server@"
-STARTUP_SCRIPT="/etc/init.d/$CUSTOMIZATION-mediaserver"
-TAR_FILE="./$DISTRIB.tar.gz"
+    exec 1<&- #< Close stdout fd.
+    exec 2<&- #< Close stderr fd.
+    exec 1<>"$LOG_FILE" #< Open stdout as $LOG_FILE for reading and writing.
+    exec 2>&1 #< Redirect stderr to stdout.
+}
 
 # Call the specified command after mounting dev, setting MNT to the mount point.
 # ATTENTION: The command is called without "exit-on-error".
@@ -51,7 +67,9 @@ copyToBootPartition()
 # [in] MNT
 copyToDataPartition()
 {
+    # Clean up sdcard from potentially unwanted files from previous installations.
     rm -rf "$MNT/$INSTALL_PATH" || true
+    rm -rf "$MNT/opt/deb"
 
     # Unpack the distro to sdcard.
     tar xfv "$TAR_FILE" -C "$MNT/" || return $?
@@ -66,28 +84,38 @@ copyToDataPartition()
     fi
 }
 
-installDeb() # package version
+# Install debs, unless the dir is missing from /opt/deb, or the proper version is installed.
+installDebs() # package [version]
 {
-    local PACKAGE="$1"
-    local VERSION="$2"
+    local -r PACKAGE="$1"; shift
+    local VERSION=""
+    if [ $# -ge 2 ]; then
+        VERSION="$2"
+    fi
 
-    local PRESENT=$(dpkg -l |grep "$PACKAGE" |grep "$VERSION" |awk '{print $3}')
-    if [ -z "$PRESENT" ]; then
-        dpkg -i "/opt/deb/$PACKAGE"/*.deb
+    local -r DEB_DIR="/opt/deb/$PACKAGE"
+    if [ ! -d "$DEB_DIR" ]; then # The deb is missing from /opt/deb.
+        return
+    fi
+
+    local -r INSTALLED_VERSION=$(dpkg -l |grep "$PACKAGE" |grep "$VERSION" |awk '{print $3}')
+    if [ -z "$INSTALLED_VERSION" ]; then
+        dpkg -i "$DEB_DIR"/*.deb
     fi
 }
 
 upgradeVms()
 {
     rm -rf "../$DISTRIB.zip" || true  #< Already unzipped, so remove .zip to save space in "/tmp".
-    rm -rf "/$MEDIASERVER_PATH/lib" "/$MEDIASERVER_PATH/bin"/core* || true
+
+    # Clean up potentially unwanted files from previous installations.
+    rm -rf "/$MEDIASERVER_PATH/lib" "/$MEDIASERVER_PATH/bin" || true
+    rm -rf "/opt/deb"
+    rm -rf "/$LITE_CLIENT_PATH"
+
     tar xfv "$TAR_FILE" -C / #< Extract the distro to the root.
 
-
-    CIFSUTILS=$(dpkg --get-selections | grep -v deinstall | grep cifs-utils | awk '{print $1}')
-    if [ -z "$CIFSUTILS" ]; then
-        dpkg -i cifs-utils/*.deb
-    fi
+    installDebs cifs-utils
 
     if [ "$BOX" = "bpi" ]; then
         # Avoid grabbing libstdc++ from mediaserver lib folder.
@@ -97,11 +125,11 @@ upgradeVms()
 
         callMounted vfat "/dev/mmcblk0p1" "/mnt/boot" copyToBootPartition
 
-        installDeb libvdpau 0.4.1
-        installDeb fontconfig 2.11
-        installDeb fonts-takao-mincho ""
-        installDeb fonts-baekmuk "" || return $?
-        installDeb fonts-arphic-ukai "" || return $?
+        installDebs libvdpau 0.4.1
+        installDebs fontconfig 2.11
+        installDebs fonts-takao-mincho
+        installDebs fonts-baekmuk
+        installDebs fonts-arphic-ukai
 
         touch "/dev/cedar_dev"
         chmod 777 "/dev/disp"
@@ -112,6 +140,8 @@ upgradeVms()
 
         /etc/init.d/nx1boot upgrade
     fi
+
+    rm -rf "/opt/deb" #< Delete deb packages to free some space.
 }
 
 getPidWhichUsesPort() # port
@@ -120,12 +150,12 @@ getPidWhichUsesPort() # port
     netstat -tpln |grep ":$PORT\s" |head -n 1 |awk '{print $NF}' |grep -o '[0-9]\+'
 }
 
-# Output nothing if the specified pid does not belong to a mediaserver; otherwise, output the pid.
+# Output nothing if pid does not belong to a mediaserver or is empty; otherwise, output the pid.
 checkMediaserverPid() # pid
 {
     local PID="$1"
 
-    if [ ! -z $(ps "$PID" |grep "/$MEDIASERVER_PATH") ]; then
+    if [ ! -z "$PID" ] && [ ! -z "$(ps "$PID" |grep "/$MEDIASERVER_PATH")" ]; then
         echo "$PID"
     fi
 }
@@ -142,29 +172,39 @@ restartMediaserver()
 
         local PID_WHICH_USES_PORT=$(getPidWhichUsesPort "$MEDIASERVER_PORT")
         local MEDIASERVER_PID=$(checkMediaserverPid "$PID_WHICH_USES_PORT")
+
         if [ ! -z "$MEDIASERVER_PID" ]; then
             echo "Upgraded mediaserver is up and running with pid $MEDIASERVER_PID at port $MEDIASERVER_PORT"
             break
         fi
 
-        echo "Another process (pid $PID_WHICH_USES_PORT) uses port $MEDIASERVER_PORT:" \
-            "killing it and restarting $CUSTOMIZATION-mediaserver"
+        if [ ! -z "$PID_WHICH_USES_PORT" ]; then
+            echo "Another process (pid $PID_WHICH_USES_PORT) uses port $MEDIASERVER_PORT:" \
+                "killing it and restarting $CUSTOMIZATION-mediaserver"
 
-        # Just in case - the mediaserver should not be running by now.
-        "$STARTUP_SCRIPT" stop || true
+            # Just in case - the mediaserver should not be running by now.
+            "$STARTUP_SCRIPT" stop || true
 
-        kill -9 "$PID_WHICH_USES_PORT" || true
+            kill -9 "$PID_WHICH_USES_PORT" || true
+        fi
     done
 }
 
 main()
 {
+    if [ $# = 0 ] || ( [ "$1" != "-v" ] && [ "$1" != "--verbose" ] ); then
+        redirectOutput "/var/log/vms-upgrade.log"
+    fi
+
+    set -x #< Log each command.
+    checkRunningUnderRoot
+    configure
     rm -rf "$FAILURE_FLAG" || true
 
-    echo "Starting VMS upgrade..."
-
+    echo "Stopping mediaserver..."
     "$STARTUP_SCRIPT" stop || true # If not stopped, try upgrading as is.
 
+    echo "Starting VMS upgrade..."
     upgradeVms
     echo "VMS upgrade succeeded"
